backend/foodgram/requirements.txt
autopep8==2.0.4
black==23.3.0
coreapi==2.3.3
django-debug-toolbar==3.2.4
django-filter==23.2
djoser==2.2.0
exceptiongroup==1.1.3
flake8-isort==6.0.0
gunicorn==21.2.0
mixer==7.1.2
pip==23.2.1
psycopg2-binary==2.9.7
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dotenv==1.0.0
reportlab==4.0.4
setuptools==58.1.0
sorl-thumbnail==12.7.0
tzdata==2023.3
